requirements.txt
pycryptodome==3.15.0
pytest==7.1.2
pytest-timeout==2.1.0

